+++ macosx/NSStringAdditions.h
@@ -36,7 +36,7 @@
 + (NSString *) stringForRatio: (CGFloat) ratio;
 
 + (NSString *) timeString: (uint64_t) seconds showSeconds: (BOOL) showSeconds;
-+ (NSString *) timeString: (NSUInteger) seconds showSeconds: (BOOL) showSeconds maxFields: (NSUInteger) max;
++ (NSString *) timeString: (uint64_t) seconds showSeconds: (BOOL) showSeconds maxFields: (NSUInteger) max;
 
 - (NSComparisonResult) compareFinder: (NSString *) string; //how the Finder compares strings
 - (NSComparisonResult) compareNumeric: (NSString *) string; //simple compare method for strings with numbers (works for IP addresses)
